requirements.txt
flask
rembg
pyrebase2